api/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-extensions==2.2.9
django-money==1.1
dropbox==10.1.1
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.0.2
filelock==3.0.12
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
identify==1.4.15
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
MarkupSafe==1.1.1
marshmallow==3.5.2
marshmallow-enum==1.5.1
mistune==0.8.4
more-itertools==8.2.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.5
nodeenv==1.3.5
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.3.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
py-moneyed==0.8.0
pycodestyle==2.5.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
Rx==1.6.1
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
stringcase==1.2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
typing-inspect==0.5.0
urllib3==1.25.9
virtualenv==20.0.19
wcwidth==0.1.9
webencodings==0.5.1
word2number==1.1
XlsxWriter==1.2.8
zipp==3.1.0
